backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-jose[cryptography]
bcrypt
pydantic[email]
slowapi
groq
reportlab
